pyproject.toml
[tool.poetry]
name = "tgpy"
version = "0.5.1"
description = "Run Python code right in your Telegram messages"
authors = ["tmat <a@tmat.me>", "vanutp <hello@vanutp.dev>", "ntonee <a12286@yandex.com>"]
license = "MIT"
documentation = "https://tgpy.tmat.me/"
repository = "https://github.com/tm-a-t/TGPy/"
readme = "README.md"
classifiers = [
    "Operating System :: OS Independent"
]

[tool.poetry.scripts]
tgpy = 'tgpy.main:main'

[tool.semantic_release]
version_variable = [
    "tgpy/version.py:__version__",
    "pyproject.toml:version"
]
branch = "master"
upload_to_repository = true
upload_to_release = true
build_command = """sed -i "s/\\(IS_DEV_BUILD *= *\\).*/\\1False/" tgpy/version.py && poetry build"""
commit_subject = 'chore(release): v{version} [skip ci]'
commit_message = ''

[tool.black]
skip-string-normalization = true
include = '/tgpy/.*\.pyi?$'

[tool.isort]
profile = "black"
src_paths = ["tgpy"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"
aiorun = "^2022.4.1"
rich = "^12.5.1"
appdirs = "^1.4.4"
telethon-v1-24 = "^1.24.2"
cryptg-anyos = "^0.4.1"

[tool.poetry.dev-dependencies]
python-semantic-release = "^7.31.2"
black = "^22.6.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
